pyproject.toml
[project]
name = "tooling-atr-experiment"
version = "0.0.1"
description = "Some experiments with ATR"
authors = [
  { name = "ASF Tooling", email = "dev@tooling.apache.org" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = "~=3.13"
dependencies = [
  "aiofiles>=24.1.0,<25.0.0",
  "aiosqlite>=0.21.0,<0.22.0",
  "alembic~=1.14",
  "asfquart", # TODO: convert asfquart from a source dependency to pypi or git dependency
  "cryptography~=44.0",
  "greenlet>=3.1.1,<4.0.0",
  "httpx~=0.27",
  "hypercorn~=0.17",
  "python-gnupg~=0.5",
  "sqlmodel~=0.0",
  "python-decouple~=3.8",
  "quart-schema~=0.21"
]

[dependency-groups]
dev = [
  "djlint>=1.36.4",
  "mypy>=1.15.0",
  "pre-commit>=2.20.0",
  "pyright>=1.1.393",
  "ruff>=0.9.4",
  "types-aiofiles>=24.1.0.20241221,<25.0.0.0",
]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24"
]

# Other configuration

# Poetry specific config

[tool.poetry]
package-mode = false

# When both project.dependencies and tool.poetry.dependencies are specified,
# project.dependencies are used for metadata when building the project,
# tool.poetry.dependencies is only used to enrich project.dependencies for locking.
# so we only need to enrich the dependencies with the source location of asfquart
[tool.poetry.dependencies]
asfquart = { path = "./asfquart", develop = true }
python = "~=3.13"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.24"

[tool.poetry.group.dev.dependencies]
djlint = "^1.36.4"
mypy = "^1.15.0"
pre-commit = ">=2.20.0"
pyright = ">=1.1.393"
ruff = ">=0.9.4"
types-aiofiles = ">=24.1.0.20241221,<25.0.0.0"
# UV specific configuration

[tool.uv.sources]
asfquart = { path = "./asfquart", editable = true }

# Additional tools

[tool.pyright]
include = ["atr"]
exclude = [
  "**/node_modules",
  "**/__pycache__",
  ".venv*",
  "asfquart",
  "tests",
  "atr/util.py"
]
ignore = []
defineConstant = { DEBUG = true }
stubPath = "typestubs"
reportMissingImports = "error"
reportMissingTypeStubs = false
pythonVersion = "3.13"
executionEnvironments = [
  { root = ".", pythonVersion = "3.13" }
]

[tool.ruff]
lint.select = [
  "I",   # isort
  "E",
  "W",
  "F",
  "N",   # pep8-naming
  "RUF", # ruff-checks
  "C90",
  "TID", # flake8-tidy-imports
  "TCH", # flake8-type-checking
  "UP"   # pyupgrade
]
lint.ignore = []
line-length = 120
exclude = ["asfquart"]

[tool.ruff.lint.per-file-ignores]
"atr/apache.py" = ["N815"]

[tool.mypy]
python_version = "3.13"
exclude = ["asfquart", "tests"]
mypy_path = "typestubs"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
# warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "asfquart.*"
ignore_errors = true

[tool.djlint]
profile = "jinja"
extension = "html"
indent = 2
blank_line_after_tag = "load,extends,include,html"
close_void_tags = true
format_css = true
format_js = true
max_line_length = 120
use_gitignore = true
preserve_blank_lines = true
ignore = "H006,H031"
include = "atr/templates"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths  = ["tests"]
